data_processing/requirements.txt
# Railway QR Tracker Data Processing Pipeline
# Smart India Hackathon 2025
# Python Dependencies

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Machine Learning Libraries
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Database Connectivity
mysql-connector-python>=8.1.0
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Async Programming
asyncio>=3.4.3
aiofiles>=23.2.0
asyncpg>=0.28.0

# Data Visualization (for reports)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Statistical Analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Configuration Management
python-dotenv>=1.0.0
pyyaml>=6.0.0
configparser>=6.0.0

# Logging and Monitoring
coloredlogs>=15.0.0
structlog>=23.1.0

# File Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Date/Time Handling
python-dateutil>=2.8.0
pytz>=2023.3

# Performance Optimization
numba>=0.57.0
joblib>=1.3.0

# API Integration
requests>=2.31.0
httpx>=0.24.0

# Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Progress Bars
tqdm>=4.66.0
rich>=13.5.0

# Memory Management
psutil>=5.9.0

# Jupyter Notebook Support (Optional)
jupyter>=1.0.0
ipython>=8.14.0

# Additional ML Libraries
imbalanced-learn>=0.11.0
hyperopt>=0.2.7

# Time Series Analysis
prophet>=1.1.0
seasonal>=0.3.0

# Text Processing (for reports)
jinja2>=3.1.0
markdown>=3.4.0

# Caching
redis>=4.6.0
diskcache>=5.6.0

# Parallel Processing
multiprocessing-logging>=0.3.4
concurrent-futures>=3.1.1

# Data Validation
cerberus>=1.3.0
marshmallow>=3.20.0

# Environment Variables
python-decouple>=3.8.0

# Cryptography (for secure connections)
cryptography>=41.0.0
